complete the function of off-line processing User and Reply data

+++ persona-core-offline/src/main/scala/org/persona/core/offline/entity/Reply.scala
@@ -2,6 +2,6 @@ package org.persona.core.offline.entity
 
 import java.sql.Date
 
-case class Reply(id: Int, replyerId: Int, anonymous: Boolean, postId: Int,
+case class Reply(id: Long, replyerId: Long, anonymous: Boolean, postId: Long,
                  countVote: Int, countComment: Int, replyTime: Date,
                  deleted: Boolean, tagAgree: Boolean, tagTop: Boolean, activeFlag: Boolean)
